hrms/requirements.txt
django-cors-headers
djangorestframework-simplejwt
psycopg2
psycopg2-binary
Pillow